[FIX] <pro>Table: Fix the positioning problem of the editor when the current record of the DataSet changes.
