+++ components-pro/button/style/index.less
@@ -109,10 +109,10 @@
 
     &::before {
       position: absolute;
-      top: -0.01rem;
-      right: -0.01rem;
-      bottom: -0.01rem;
-      left: -0.01rem;
+      top: 0;
+      right: 0;
+      bottom: 0;
+      left: 0;
       z-index: 1;
       background: #fff;
       border-radius: inherit;
